LayoutTests/imported/w3c/web-platform-tests/user-timing/user_timing_exists.any.worker-expected.txt


PASS self.performance.mark is defined.
PASS self.performance.clearMarks is defined.
PASS self.performance.measure is defined.
PASS self.performance.clearMeasures is defined.
